global/generators/requirements.txt
# VWE Code Generators Requirements
# Python dependencies for the code generators

# Core dependencies
pathlib2>=2.3.7; python_version < "3.4"
typing-extensions>=4.0.0

# Template processing (if needed for future enhancements)
jinja2>=3.1.0
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: For advanced template processing
# cookiecutter>=2.1.0
# click>=8.0.0
